requirements.txt
djangorestframework==3.15.2
gliner==0.2.13
Django==5.1.2
django-prometheus==2.3.1
